Added a configuration class and Credit/Round state tracker. Changed slot machine implementations to use these new classes. Now can state current PBR and HR.

+++ SlotsProto/CombinationEvaluator.h
@@ -6,7 +6,7 @@ class CombinationEvaluator
 public:
    CombinationEvaluator();
    CombinationEvaluator(int);
-   int checkPayLineCombinations(const std::vector<SymbolCombination>&);
+   size_t checkPayLineCombinations(const std::vector<SymbolCombination>&);
 private:
    std::map<SymbolCombination, int> evalmap; //Should change bool to int/double(amount you would win)
 };
